Django_progects/django_shop/requirements.txt
Django==2.2.4
django-bootstrap4==0.0.8
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.10.2
easy-thumbnails==2.6
Pillow==6.1.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
